fix: relationships-always-being-setup
